+++ src/Pages/Login/SignInForm/components/index.ts
@@ -1,0 +1,6 @@
+export { PasswordForm } from './PasswordForm';
+export { AddressForm } from './AddressForm';
+export { SignInPanel } from './Panel';
+export { NameForm } from './NameForm';
+export { SignInInput } from './SinginInput';
+export { EmailForm } from './EmailForm';
